+++ crd-generator/api-v2/src/main/java/io/fabric8/crdv2/generator/AbstractJsonSchema.java
@@ -85,7 +85,6 @@ import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static io.fabric8.crdv2.generator.CRDUtils.toTargetType;
 import static java.util.Optional.ofNullable;
 
 /**
@@ -232,7 +231,7 @@ public abstract class AbstractJsonSchema<T extends KubernetesJSONSchemaProps, V 
 
     private boolean required;
     private final String description;
-    private final Object defaultValue;
+    private final JsonNode defaultValue;
     private Double min;
     private Boolean exclusiveMinimum;
     private Double max;
@@ -314,14 +313,27 @@ public abstract class AbstractJsonSchema<T extends KubernetesJSONSchemaProps, V 
 
       // TODO: should the following be deprecated?
       required = beanProperty.getAnnotation(Required.class) != null;
+      defaultValue = toDefault(beanProperty);
+    }
 
-      if (beanProperty.getMetadata().getDefaultValue() != null) {
-        defaultValue = toTargetType(beanProperty.getType(), beanProperty.getMetadata().getDefaultValue());
-      } else if (ofNullable(beanProperty.getAnnotation(Default.class)).map(Default::value).isPresent()) {
-        defaultValue = toTargetType(beanProperty.getType(),
-            ofNullable(beanProperty.getAnnotation(Default.class)).map(Default::value).get());
-      } else {
-        defaultValue = null;
+    JsonNode toDefault(BeanProperty beanProperty) {
+      Optional<String> defaultAnnotationValue = ofNullable(beanProperty.getAnnotation(Default.class)).map(Default::value);
+      String value = defaultAnnotationValue.orElse(beanProperty.getMetadata().getDefaultValue());
+
+      if (value == null) {
+        return null;
+      }
+      Optional<Class<?>> rawType = Optional.ofNullable(beanProperty.getType()).map(JavaType::getRawClass);
+      try {
+        Object typedValue = resolvingContext.kubernetesSerialization.unmarshal(value, rawType.orElse(Object.class));
+        return resolvingContext.kubernetesSerialization.convertValue(typedValue, JsonNode.class);
+      } catch (Exception e) {
+        if (defaultAnnotationValue.isEmpty()) {
+          LOGGER.warn("Cannot parse default value: '" + value
+              + "' from JsonProperty annotation as valid YAML or JSON, no default value will be used.");
+          return null;
+        }
+        throw new IllegalArgumentException("Cannot parse default value: '" + value + "' as valid YAML or JSON.", e);
       }
     }
 
@@ -355,14 +367,7 @@ public abstract class AbstractJsonSchema<T extends KubernetesJSONSchemaProps, V 
 
     public void updateSchema(T schema) {
       schema.setDescription(description);
-
-      if (defaultValue != null) {
-        try {
-          schema.setDefault(resolvingContext.kubernetesSerialization.convertValue(defaultValue, JsonNode.class));
-        } catch (IllegalArgumentException e) {
-          throw new IllegalArgumentException("Cannot parse default value: '" + defaultValue + "' as valid YAML.", e);
-        }
-      }
+      schema.setDefault(defaultValue);
       if (nullable) {
         schema.setNullable(true);
       }