+++ crd-generator/api-v2/src/main/java/io/fabric8/crdv2/generator/CRDUtils.java
@@ -16,12 +16,10 @@
 package io.fabric8.crdv2.generator;
 
 import com.fasterxml.jackson.databind.BeanDescription;
-import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationConfig;
 import com.fasterxml.jackson.databind.introspect.BeanPropertyDefinition;
 
-import java.text.NumberFormat;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -94,23 +92,4 @@ public class CRDUtils {
     return res;
   }
 
-  static Object toTargetType(JavaType type, String value) {
-    if (type == null || value == null) {
-      return null;
-    }
-    try {
-      if (Number.class.isAssignableFrom(type.getRawClass()) || int.class.isAssignableFrom(type.getRawClass())
-          || long.class.isAssignableFrom(type.getRawClass()) || float.class.isAssignableFrom(type.getRawClass())
-          || double.class.isAssignableFrom(type.getRawClass())) {
-        return NumberFormat.getInstance().parse(value);
-      }
-      if (Boolean.class.isAssignableFrom(type.getRawClass()) || boolean.class.isAssignableFrom(type.getRawClass())) {
-        return Boolean.valueOf(value);
-      }
-    } catch (Exception ex) {
-      // NO OP
-    }
-    return value;
-
-  }
 }