+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Bugs
 * Fix #6892: rolling().restart() doesn't remove preexistent pod template annotations
+* Fix #6908: The Default annotation and JsonProperty default value should accept JSON values
 * Fix #6906: Knative VolatileTime should be serialized as String
 * Fix #6930: Add support for Boolean enums in the java-generator
 * Fix #6886: Remove invalid JUnit 4 references